+++ template-vanilla-js/src/popup/index.js
@@ -1,0 +1,18 @@
+import './index.css'
+
+const crx = 'create-chrome-ext'
+
+document.querySelector('#app').innerHTML = `
+<main>
+<h3>Popup Page!</h3>
+
+<h6>v 0.0.1</h6>
+
+<a
+  href="https://github.com/guocaoyi/create-chrome-ext"
+  target="_blank"
+>
+  Power by ${crx}
+</a>
+</main>
+`