+++ template-vue-ts/src/newtab/index.ts
@@ -1,5 +1,0 @@
-import { createApp } from 'vue'
-import './style.css'
-import App from './NewTab.vue'
-
-createApp(App).mount('#app')