+++ template-vue-ts/src/popup/index.ts
@@ -1,5 +1,0 @@
-import { createApp } from 'vue'
-import './style.css'
-import App from './Popup.vue'
-
-createApp(App).mount('#app')